+++ main.go
@@ -287,6 +287,12 @@ func main() {
 		"div": func(a, b int) int {
 			return a / b
 		},
+		"add": func(a, b int) int {
+			return a + b
+		},
+		"sub": func(a, b int) int {
+			return a - b
+		},
 		"toInt": func(value interface{}) int {
 			switch v := value.(type) {
 			case string: